Database connected with Ado.Net - added Multilayer arhitekture

+++ WebCRUD.API/Models/ProductModel.cs
@@ -7,8 +7,9 @@ namespace WebCRUD.API.Models
 {
     public class ProductModel
     {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public double Price { get; set; }
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public int Price { get; set; }
+        
     }
 }
